+++ source/Core/CastorUtils/Src/NonAlignedMemoryAllocator.hpp
@@ -44,7 +44,7 @@ namespace Castor
 		 *\brief		Alloue de la mémoire.
 		 *\return		L'adresse mémoire de l'espace alloué.
 		 */
-		static uint8_t * Allocate( uint32_t p_size )
+		static uint8_t * Allocate( size_t p_size )
 		{
 			return new uint8_t[p_size];;
 		}
